更新navbar按鈕
