+++ components/layout/default-layout/navbar.module.scss
@@ -1,0 +1,49 @@
+$primary: #013c64;
+$secondary: #265475;
+$orange: #ff9720;
+$gray: #858585;
+$light-gray: #d9d9d9;
+$super-light-gray: #f5f5f5;
+
+.navbar {
+  background-color: $primary;
+  padding-inline: 30px;
+  padding-block: 10px;
+  h1 {
+    color: #fff;
+    font-weight: bolder;
+    margin-bottom: 0;
+  }
+  ul {
+    margin-bottom: 0;
+    li {
+      margin-left: 20px;
+      a {
+        color: #fff;
+        display: block;
+        text-align: center;
+        padding-block: 8px;
+      }
+      a:hover {
+        color: $orange;
+        border-bottom: 2px solid $orange;
+      }
+      .loginBtn {
+        background-color: #fff;
+        color: $orange;
+        padding-inline: 12px;
+        border-radius: 20px;
+      }
+      .loginBtn:hover {
+        background-color: $orange;
+        color: #fff;
+      }
+      .cartIcon {
+        font-size: 18px;
+      }
+      .cartBtn:hover {
+        border-bottom: none;
+      }
+    }
+  }
+}
