+++ pages/home/index.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import Navbar from '@/components/layout/default-layout/navbar'
+
+export default function Index() {
+  return (
+    <>
+      <Navbar />
+    </>
+  )
+}
